juanstreamlit/requirements.txt
certifi==2024.2.2
firebase_admin==6.5.0
folium==0.16.0
geopy==2.4.1
googletrans==4.0.0rc1
ipython==8.12.3
Kivy==2.3.0
openpyxl==3.1.2
pandas==2.2.2
protobuf==3.20.3
PyPDF2==1.28.6
PyPDF2==3.0.1
Requests==2.32.3
st_circular_progress==0.3.9
streamlit==1.36.0
tkcalendar==1.6.1
xmltodict==0.13.0
